+++ maven/src/main/java/org/owasp/dependencycheck/maven/BaseDependencyCheckMojo.java
@@ -17,6 +17,8 @@
  */
 package org.owasp.dependencycheck.maven;
 
+import com.github.packageurl.MalformedPackageURLException;
+import com.github.packageurl.PackageURL.StandardTypes;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.repository.ArtifactRepository;
 import org.apache.maven.artifact.versioning.ArtifactVersion;
@@ -36,14 +38,12 @@ import org.apache.maven.reporting.MavenReportException;
 import org.apache.maven.settings.Proxy;
 import org.apache.maven.settings.Server;
 import org.apache.maven.shared.transfer.artifact.ArtifactCoordinate;
-import org.apache.maven.shared.transfer.artifact.DefaultArtifactCoordinate;
 import org.apache.maven.shared.transfer.artifact.TransferUtils;
 import org.apache.maven.shared.transfer.artifact.resolve.ArtifactResolver;
 import org.apache.maven.shared.transfer.artifact.resolve.ArtifactResolverException;
 import org.apache.maven.shared.dependency.graph.DependencyGraphBuilder;
 import org.apache.maven.shared.dependency.graph.DependencyGraphBuilderException;
 import org.apache.maven.shared.dependency.graph.DependencyNode;
-import org.apache.maven.shared.dependency.graph.internal.DefaultDependencyNode;
 import org.apache.maven.shared.model.fileset.FileSet;
 import org.apache.maven.shared.model.fileset.util.FileSetManager;
 import org.owasp.dependencycheck.Engine;
@@ -53,7 +53,6 @@ import org.owasp.dependencycheck.data.nvdcve.DatabaseException;
 import org.owasp.dependencycheck.dependency.Confidence;
 import org.owasp.dependencycheck.dependency.Dependency;
 import org.owasp.dependencycheck.dependency.EvidenceType;
-import org.owasp.dependencycheck.dependency.Identifier;
 import org.owasp.dependencycheck.dependency.Vulnerability;
 import org.owasp.dependencycheck.exception.DependencyNotFoundException;
 import org.owasp.dependencycheck.exception.ExceptionCollection;
@@ -77,6 +76,10 @@ import java.util.Locale;
 import org.apache.maven.artifact.resolver.filter.ArtifactFilter;
 import org.apache.maven.artifact.resolver.filter.ExcludesArtifactFilter;
 import org.apache.maven.shared.dependency.graph.traversal.CollectingDependencyNodeVisitor;
+import org.owasp.dependencycheck.agent.DependencyCheckScanAgent;
+import org.owasp.dependencycheck.dependency.naming.GenericIdentifier;
+import org.owasp.dependencycheck.dependency.naming.Identifier;
+import org.owasp.dependencycheck.dependency.naming.PurlIdentifier;
 
 /**
  * @author Jeremy Long
@@ -104,12 +107,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     // <editor-fold defaultstate="collapsed" desc="Maven bound parameters and components">
     /**
      * Sets whether or not the external report format should be used.
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "metaFileName", defaultValue = "dependency-check.ser", required = true)
-    private String dataFileName;
-    /**
-     * Sets whether or not the build should be interrupted because of a failure.
      */
     @SuppressWarnings("CanBeFinal")
     @Parameter(property = "failOnError", defaultValue = "true", required = true)
@@ -202,21 +199,7 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     @SuppressWarnings("CanBeFinal")
     @Parameter(property = "enableRetired")
     private Boolean enableRetired;
-    /**
-     * Generate aggregate reports in multi-module projects.
-     *
-     * @deprecated use the aggregate goal instead
-     */
-    @Parameter(property = "aggregate")
-    @Deprecated
-    private Boolean aggregate;
-    /**
-     * Use pom dependency information for snapshot dependencies that are part of
-     * the Maven reactor while aggregate scanning a multi-module project.
-     *
-     */
-    @Parameter(property = "dependency-check.virtualSnapshotsFromReactor", defaultValue = "true")
-    private Boolean virtualSnapshotsFromReactor;
+
     /**
      * The report format to be generated (HTML, XML, VULN, ALL). This
      * configuration option has no affect if using this within the Site plug-in
@@ -345,28 +328,16 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     @Parameter(property = "nodeAuditAnalyzerEnabled", required = false)
     private Boolean nodeAuditAnalyzerEnabled;
     /**
-     * Sets whether or not the Node Security Project Analyzer should be used.
-     *
-     * @deprecated As of release 3.3.3, replaced by
-     * {@link #nodeAuditAnalyzerEnabled}
-     */
-    @Deprecated
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "nspAnalyzerEnabled", required = false)
-    private Boolean nspAnalyzerEnabled;
-    /**
      * Sets whether or not the Retirejs Analyzer should be used.
      */
     @SuppressWarnings("CanBeFinal")
     @Parameter(property = "retireJsAnalyzerEnabled", required = false)
     private Boolean retireJsAnalyzerEnabled;
-
     /**
      * Whether or not the .NET Assembly Analyzer is enabled.
      */
     @Parameter(property = "assemblyAnalyzerEnabled", required = false)
     private Boolean assemblyAnalyzerEnabled;
-
     /**
      * Whether or not the .NET Nuspec Analyzer is enabled.
      */
@@ -480,15 +451,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     @Parameter(property = "nexusUrl", required = false)
     private String nexusUrl;
     /**
-     * The id of a server defined in the settings.xml that configures the
-     * credentials (username and password) for a Nexus server's REST API end
-     * point. When not specified the communication with the Nexus server's REST
-     * API will be unauthenticated.
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "nexusServerId", required = false)
-    private String nexusServerId;
-    /**
      * Whether or not the configured proxy is used to connect to Nexus.
      */
     @SuppressWarnings("CanBeFinal")
@@ -581,13 +543,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     private boolean skipSystemScope = false;
 
     /**
-     * Skip Analysis for dependencyManagement section.
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "skipDependencyManagement", defaultValue = "true", required = false)
-    private boolean skipDependencyManagement = true;
-
-    /**
      * Skip analysis for dependencies which type matches this regular
      * expression.
      */
@@ -605,26 +560,14 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      * Data Mirror URL for CVE 1.2.
      */
     @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "cveUrl12Modified", defaultValue = "", required = false)
-    private String cveUrl12Modified;
-    /**
-     * Data Mirror URL for CVE 2.0.
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "cveUrl20Modified", defaultValue = "", required = false)
-    private String cveUrl20Modified;
+    @Parameter(property = "cveUrlModified", defaultValue = "", required = false)
+    private String cveUrlModified;
     /**
      * Base Data Mirror URL for CVE 1.2.
      */
     @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "cveUrl12Base", defaultValue = "", required = false)
-    private String cveUrl12Base;
-    /**
-     * Data Mirror URL for CVE 2.0.
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "cveUrl20Base", defaultValue = "", required = false)
-    private String cveUrl20Base;
+    @Parameter(property = "cveUrlBase", defaultValue = "", required = false)
+    private String cveUrlBase;
     /**
      * Optionally skip excessive CVE update checks for a designated duration in
      * hours.
@@ -660,25 +603,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     private Retirejs retirejs;
 
     /**
-     * The Proxy URL.
-     *
-     * @deprecated Please use mavenSettings instead
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "proxyUrl", defaultValue = "", required = false)
-    @Deprecated
-    private String proxyUrl = null;
-    /**
-     * Sets whether or not the external report format should be used.
-     *
-     * @deprecated the internal report is no longer supported
-     */
-    @SuppressWarnings("CanBeFinal")
-    @Parameter(property = "externalReport")
-    @Deprecated
-    private String externalReport = null;
-
-    /**
      * The artifact scope filter.
      */
     private Filter<String> artifactScopeExcluded;
@@ -708,9 +632,9 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      * @return true if the groupId, artifactId, and version match
      */
     private static boolean artifactsMatch(org.apache.maven.model.Dependency d, Artifact a) {
-        return (isEqualOrNull(a.getArtifactId(), d.getArtifactId()))
-                && (isEqualOrNull(a.getGroupId(), d.getGroupId()))
-                && (isEqualOrNull(a.getVersion(), d.getVersion()));
+        return isEqualOrNull(a.getArtifactId(), d.getArtifactId())
+                && isEqualOrNull(a.getGroupId(), d.getGroupId())
+                && isEqualOrNull(a.getVersion(), d.getVersion());
     }
 
     /**
@@ -740,24 +664,9 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         if (shouldSkip) {
             getLog().info("Skipping " + getName(Locale.US));
         } else {
-            validateAggregate();
-            project.setContextValue(getOutputDirectoryContextKey(), this.outputDirectory);
+            //TODO remove
+            //project.setContextValue(getOutputDirectoryContextKey(), this.outputDirectory);
             runCheck();
-        }
-    }
-
-    /**
-     * Checks if the aggregate configuration parameter has been set to true. If
-     * it has a MojoExecutionException is thrown because the aggregate
-     * configuration parameter is no longer supported.
-     *
-     * @throws MojoExecutionException thrown if aggregate is set to true
-     */
-    private void validateAggregate() throws MojoExecutionException {
-        if (aggregate != null && aggregate) {
-            final String msg = "Aggregate configuration detected - as of dependency-check 1.2.8 this no longer supported. "
-                    + "Please use the aggregate goal instead.";
-            throw new MojoExecutionException(msg);
         }
     }
 
@@ -819,12 +728,8 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         }
 
         generatingSite = true;
-        try {
-            validateAggregate();
-        } catch (MojoExecutionException ex) {
-            throw new MavenReportException(ex.getMessage());
-        }
-        project.setContextValue(getOutputDirectoryContextKey(), getReportOutputDirectory());
+        //TODO remove
+        //project.setContextValue(getOutputDirectoryContextKey(), getReportOutputDirectory());
         try {
             runCheck();
         } catch (MojoExecutionException ex) {
@@ -854,10 +759,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      * @return the directory to write the report(s)
      */
     protected File getCorrectOutputDirectory(MavenProject current) {
-        final Object obj = current.getContextValue(getOutputDirectoryContextKey());
-        if (obj != null && obj instanceof File) {
-            return (File) obj;
-        }
         File target = new File(current.getBuild().getDirectory());
         if (target.getParentFile() != null && "target".equals(target.getParentFile().getName())) {
             target = target.getParentFile();
@@ -917,68 +818,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
     }
 
     /**
-     * Converts the dependency into a dependency node
-     *
-     * @param buildingRequest the Maven building request
-     * @param parent the parent node
-     * @param dependency the dependency to convert
-     * @return the resulting dependency node
-     * @throws ArtifactResolverException thrown if there is an error during
-     * conversion
-     */
-    private DependencyNode toDependencyNode(ProjectBuildingRequest buildingRequest, DependencyNode parent,
-            org.apache.maven.model.Dependency dependency) throws ArtifactResolverException {
-
-        final DefaultArtifactCoordinate coordinate = new DefaultArtifactCoordinate();
-
-        coordinate.setGroupId(dependency.getGroupId());
-        coordinate.setArtifactId(dependency.getArtifactId());
-        coordinate.setVersion(dependency.getVersion());
-        coordinate.setExtension(dependency.getType());
-        coordinate.setClassifier(dependency.getClassifier());
-
-        final Artifact artifact = artifactResolver.resolveArtifact(buildingRequest, coordinate).getArtifact();
-
-        artifact.setScope(dependency.getScope());
-
-        return new DefaultDependencyNode(parent, artifact, dependency.getVersion(), dependency.getScope(), null);
-
-    }
-
-    /**
-     * Collects the dependencies defined in the dependency management
-     * configuration.
-     *
-     * @param buildingRequest the Maven building request
-     * @param project the Maven project
-     * @param nodes the collection to add the dependencies to from the
-     * dependency management configuration
-     * @param aggregate whether or not this is an aggregate build
-     * @return a collection of exceptions; if the return value is null no
-     * exceptions occurred
-     */
-    private ExceptionCollection collectDependencyManagementDependencies(ProjectBuildingRequest buildingRequest, MavenProject project,
-            List<DependencyNode> nodes, boolean aggregate) {
-        if (skipDependencyManagement || project.getDependencyManagement() == null) {
-            return null;
-        }
-
-        ExceptionCollection exCol = null;
-        for (org.apache.maven.model.Dependency dependency : project.getDependencyManagement().getDependencies()) {
-            try {
-                nodes.add(toDependencyNode(buildingRequest, null, dependency));
-            } catch (ArtifactResolverException ex) {
-                getLog().debug(String.format("Aggregate : %s", aggregate));
-                if (exCol == null) {
-                    exCol = new ExceptionCollection();
-                }
-                exCol.addException(ex);
-            }
-        }
-        return exCol;
-    }
-
-    /**
      * Resolves the projects artifacts using Aether and scans the resulting
      * dependencies.
      *
@@ -993,7 +832,7 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      */
     private ExceptionCollection collectMavenDependencies(Engine engine, MavenProject project,
             List<DependencyNode> nodes, ProjectBuildingRequest buildingRequest, boolean aggregate) {
-        ExceptionCollection exCol = collectDependencyManagementDependencies(buildingRequest, project, nodes, aggregate);
+        ExceptionCollection exCol = null;
         for (DependencyNode dependencyNode : nodes) {
             if (artifactScopeExcluded.passes(dependencyNode.getArtifact().getScope())
                     || artifactTypeExcluded.passes(dependencyNode.getArtifact().getType())) {
@@ -1045,11 +884,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
                         }
                         exCol.addException(ex);
                     }
-                    continue;
-                }
-                if (aggregate && virtualSnapshotsFromReactor
-                        && dependencyNode.getArtifact().isSnapshot() && aggregate
-                        && addSnapshotReactorDependency(engine, dependencyNode.getArtifact())) {
                     continue;
                 }
                 isResolved = result.isResolved();
@@ -1211,24 +1045,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      * <code>false</code>
      */
     private boolean addReactorDependency(Engine engine, Artifact artifact) {
-        return addVirtualDependencyFromReactor(engine, artifact, "Unable to resolve %s as it has not been built yet "
-                + "- creating a virtual dependency instead.");
-    }
-
-    /**
-     * Checks if the current artifact is actually in the reactor projects. If
-     * true a virtual dependency is created based on the evidence in the
-     * project.
-     *
-     * @param engine a reference to the engine being used to scan
-     * @param artifact the artifact being analyzed in the mojo
-     * @param infoLogTemplate the template for the infoLog entry written when a
-     * virtual dependency is added. Needs a single %s placeholder for the
-     * location of the displayName in the message
-     * @return <code>true</code> if the artifact is in the reactor; otherwise
-     * <code>false</code>
-     */
-    private boolean addVirtualDependencyFromReactor(Engine engine, Artifact artifact, String infoLogTemplate) {
 
         getLog().debug(String.format("Checking the reactor projects (%d) for %s:%s:%s",
                 reactorProjects.size(),
@@ -1246,7 +1062,7 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
 
                 final String displayName = String.format("%s:%s:%s",
                         prj.getGroupId(), prj.getArtifactId(), prj.getVersion());
-                getLog().info(String.format(infoLogTemplate,
+                getLog().info(String.format("Unable to resolve %s as it has not been built yet - creating a virtual dependency instead.",
                         displayName));
                 final File pom = new File(prj.getBasedir(), "pom.xml");
                 final Dependency d;
@@ -1269,11 +1085,15 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
                 d.addEvidence(EvidenceType.VENDOR, "project", "groupid", prj.getGroupId(), Confidence.HIGHEST);
                 d.addEvidence(EvidenceType.PRODUCT, "project", "groupid", prj.getGroupId(), Confidence.LOW);
                 d.setEcosystem(JarAnalyzer.DEPENDENCY_ECOSYSTEM);
-                final Identifier id = new Identifier();
-                id.setType("maven");
-                id.setConfidence(Confidence.HIGHEST);
-                id.setValue(displayName);
-                d.addIdentifier(id);
+                Identifier id;
+                try {
+                    id = new PurlIdentifier(StandardTypes.MAVEN, artifact.getGroupId(),
+                            artifact.getArtifactId(), artifact.getVersion(), Confidence.HIGHEST);
+                } catch (MalformedPackageURLException ex) {
+                    getLog().debug("Unable to create PackageURL object:" + key);
+                    id = new GenericIdentifier("maven:" + key, Confidence.HIGHEST);
+                }
+                d.addSoftwareIdentifier(id);
                 //TODO unify the setName/version and package path - they are equivelent ideas submitted by two seperate committers
                 d.setName(String.format("%s:%s", prj.getGroupId(), prj.getArtifactId()));
                 d.setVersion(prj.getVersion());
@@ -1300,24 +1120,6 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
             }
         }
         return false;
-    }
-
-    /**
-     * Checks if the current artifact is actually in the reactor projects. If
-     * true a virtual dependency is created based on the evidence in the
-     * project.
-     *
-     * @param engine a reference to the engine being used to scan
-     * @param artifact the artifact being analyzed in the mojo
-     * @return <code>true</code> if the artifact is a snapshot artifact in the
-     * reactor; otherwise <code>false</code>
-     */
-    private boolean addSnapshotReactorDependency(Engine engine, Artifact artifact) {
-        if (!artifact.isSnapshot()) {
-            return false;
-        }
-        return addVirtualDependencyFromReactor(engine, artifact, "Found snapshot reactor project in aggregate for %s - "
-                + "creating a virtual dependency as the snapshot found in the repository may contain outdated dependencies.");
     }
 
     /**
@@ -1552,15 +1354,10 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
             }
         }
         settings.setBooleanIfNotNull(Settings.KEYS.AUTO_UPDATE, autoUpdate);
+
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_EXPERIMENTAL_ENABLED, enableExperimental);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_RETIRED_ENABLED, enableRetired);
-        if (externalReport != null) {
-            getLog().warn("The 'externalReport' option was set; this configuration option has been removed. "
-                    + "Please update the dependency-check-maven plugin's configuration");
-        }
-        if (proxyUrl != null && !proxyUrl.isEmpty()) {
-            getLog().warn("Deprecated configuration detected, proxyUrl will be ignored; use the maven settings to configure the proxy instead");
-        }
+
         final Proxy proxy = getMavenProxy();
         if (proxy != null) {
             settings.setString(Settings.KEYS.PROXY_SERVER, proxy.getHost());
@@ -1573,9 +1370,11 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         }
         final String[] suppressions = determineSuppressions();
         settings.setArrayIfNotEmpty(Settings.KEYS.SUPPRESSION_FILE, suppressions);
+
         settings.setBooleanIfNotNull(Settings.KEYS.UPDATE_VERSION_CHECK_ENABLED, versionCheckEnabled);
         settings.setStringIfNotEmpty(Settings.KEYS.CONNECTION_TIMEOUT, connectionTimeout);
         settings.setStringIfNotEmpty(Settings.KEYS.HINTS_FILE, hintsFile);
+
         //File Type Analyzer Settings
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_JAR_ENABLED, jarAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_NUSPEC_ENABLED, nuspecAnalyzerEnabled);
@@ -1587,42 +1386,14 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_ARCHIVE_ENABLED, archiveAnalyzerEnabled);
         settings.setStringIfNotEmpty(Settings.KEYS.ADDITIONAL_ZIP_EXTENSIONS, zipExtensions);
         settings.setStringIfNotEmpty(Settings.KEYS.ANALYZER_ASSEMBLY_MONO_PATH, pathToMono);
+
         settings.setStringIfNotEmpty(Settings.KEYS.ANALYZER_NEXUS_URL, nexusUrl);
-        if (nexusServerId != null) {
-            final Server server = settingsXml.getServer(nexusServerId);
-            if (server != null) {
-                final String nexusUser = server.getUsername();
-                String nexusPassword = null;
-                try {
-                    nexusPassword = decryptServerPassword(server);
-                } catch (SecDispatcherException ex) {
-                    if (ex.getCause() instanceof FileNotFoundException
-                            || (ex.getCause() != null && ex.getCause().getCause() instanceof FileNotFoundException)) {
-                        //maybe its not encrypted?
-                        final String tmp = server.getPassword();
-                        if (tmp.startsWith("{") && tmp.endsWith("}")) {
-                            getLog().error(String.format(
-                                    "Unable to decrypt the server password for server id '%s' in settings.xml%n\tCause: %s",
-                                    serverId, ex.getMessage()));
-                        } else {
-                            nexusPassword = tmp;
-                        }
-                    } else {
-                        getLog().error(String.format(
-                                "Unable to decrypt the server password for server id '%s' in settings.xml%n\tCause: %s",
-                                serverId, ex.getMessage()));
-                    }
-                }
-                settings.setStringIfNotEmpty(Settings.KEYS.ANALYZER_NEXUS_USER, nexusUser);
-                settings.setStringIfNotEmpty(Settings.KEYS.ANALYZER_NEXUS_PASSWORD, nexusPassword);
-            } else {
-                getLog().error(String.format("Server '%s' not found in the settings.xml file", serverId));
-            }
-        }
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_NEXUS_USES_PROXY, nexusUsesProxy);
+
         settings.setStringIfNotNull(Settings.KEYS.ANALYZER_ARTIFACTORY_URL, artifactoryAnalyzerUrl);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_ARTIFACTORY_USES_PROXY, artifactoryAnalyzerUseProxy);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_ARTIFACTORY_PARALLEL_ANALYSIS, artifactoryAnalyzerParallelAnalysis);
+
         if (Boolean.TRUE.equals(artifactoryAnalyzerEnabled)) {
             if (artifactoryAnalyzerServerId != null) {
                 final Server server = settingsXml.getServer(artifactoryAnalyzerServerId);
@@ -1636,6 +1407,7 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
             }
             settings.setStringIfNotNull(Settings.KEYS.ANALYZER_ARTIFACTORY_BEARER_TOKEN, artifactoryAnalyzerBearerToken);
         }
+
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_PYTHON_DISTRIBUTION_ENABLED, pyDistributionAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_PYTHON_PACKAGE_ENABLED, pyPackageAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_RUBY_GEMSPEC_ENABLED, rubygemsAnalyzerEnabled);
@@ -1644,31 +1416,42 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_AUTOCONF_ENABLED, autoconfAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_COMPOSER_LOCK_ENABLED, composerAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_NODE_PACKAGE_ENABLED, nodeAnalyzerEnabled);
-        if (nspAnalyzerEnabled != null) {
-            getLog().error("The nspAnalyzerEnabled configuration has been deprecated and replaced by nodeAuditAnalyzerEnabled");
-            getLog().error("The nspAnalyzerEnabled configuration will be removed in the next major release");
-            settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_NODE_AUDIT_ENABLED, nspAnalyzerEnabled);
-        }
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_NODE_AUDIT_ENABLED, nodeAuditAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_RETIREJS_ENABLED, retireJsAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_BUNDLE_AUDIT_ENABLED, bundleAuditAnalyzerEnabled);
         settings.setStringIfNotNull(Settings.KEYS.ANALYZER_BUNDLE_AUDIT_PATH, bundleAuditPath);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_COCOAPODS_ENABLED, cocoapodsAnalyzerEnabled);
         settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_SWIFT_PACKAGE_MANAGER_ENABLED, swiftPackageManagerAnalyzerEnabled);
+
         if (retirejs != null) {
             settings.setBooleanIfNotNull(Settings.KEYS.ANALYZER_RETIREJS_FILTER_NON_VULNERABLE, retirejs.getFilterNonVulnerable());
             settings.setArrayIfNotEmpty(Settings.KEYS.ANALYZER_RETIREJS_FILTERS, retirejs.getFilters());
         }
+
         //Database configuration
         settings.setStringIfNotEmpty(Settings.KEYS.DB_DRIVER_NAME, databaseDriverName);
         settings.setStringIfNotEmpty(Settings.KEYS.DB_DRIVER_PATH, databaseDriverPath);
         settings.setStringIfNotEmpty(Settings.KEYS.DB_CONNECTION_STRING, connectionString);
+
         if (databaseUser == null && databasePassword == null && serverId != null) {
             final Server server = settingsXml.getServer(serverId);
             if (server != null) {
                 databaseUser = server.getUsername();
                 try {
-                    databasePassword = decryptServerPassword(server);
+                    //CSOFF: LineLength
+                    //The following fix was copied from:
+                    //   https://github.com/bsorrentino/maven-confluence-plugin/blob/master/maven-confluence-reporting-plugin/src/main/java/org/bsc/maven/confluence/plugin/AbstractBaseConfluenceMojo.java
+                    //
+                    // FIX to resolve
+                    // org.sonatype.plexus.components.sec.dispatcher.SecDispatcherException:
+                    // java.io.FileNotFoundException: ~/.settings-security.xml (No such file or directory)
+                    //
+                    //CSON: LineLength
+                    if (securityDispatcher instanceof DefaultSecDispatcher) {
+                        ((DefaultSecDispatcher) securityDispatcher).setConfigurationFile("~/.m2/settings-security.xml");
+                    }
+
+                    databasePassword = securityDispatcher.decrypt(server.getPassword());
                 } catch (SecDispatcherException ex) {
                     if (ex.getCause() instanceof FileNotFoundException
                             || (ex.getCause() != null && ex.getCause().getCause() instanceof FileNotFoundException)) {
@@ -1691,41 +1474,17 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
                 getLog().error(String.format("Server '%s' not found in the settings.xml file", serverId));
             }
         }
+
         settings.setStringIfNotEmpty(Settings.KEYS.DB_USER, databaseUser);
         settings.setStringIfNotEmpty(Settings.KEYS.DB_PASSWORD, databasePassword);
         settings.setStringIfNotEmpty(Settings.KEYS.DATA_DIRECTORY, dataDirectory);
-        settings.setStringIfNotEmpty(Settings.KEYS.CVE_MODIFIED_12_URL, cveUrl12Modified);
-        settings.setStringIfNotEmpty(Settings.KEYS.CVE_MODIFIED_20_URL, cveUrl20Modified);
-        settings.setStringIfNotEmpty(Settings.KEYS.CVE_SCHEMA_1_2, cveUrl12Base);
-        settings.setStringIfNotEmpty(Settings.KEYS.CVE_SCHEMA_2_0, cveUrl20Base);
+
+        settings.setStringIfNotEmpty(Settings.KEYS.CVE_MODIFIED_JSON, cveUrlModified);
+        settings.setStringIfNotEmpty(Settings.KEYS.CVE_BASE_JSON, cveUrlBase);
         settings.setIntIfNotNull(Settings.KEYS.CVE_CHECK_VALID_FOR_HOURS, cveValidForHours);
+
         artifactScopeExcluded = new ArtifactScopeExcluded(skipTestScope, skipProvidedScope, skipSystemScope, skipRuntimeScope);
         artifactTypeExcluded = new ArtifactTypeExcluded(skipArtifactType);
-    }
-
-    /**
-     * Obtains the configured password for the given server.
-     *
-     * @param server the configured server from the settings.xml
-     * @return the decrypted password from the Maven configuration
-     * @throws SecDispatcherException thrown if there is an error decrypting the
-     * password
-     */
-    private String decryptServerPassword(Server server) throws SecDispatcherException {
-        //CSOFF: LineLength
-        //The following fix was copied from:
-        //   https://github.com/bsorrentino/maven-confluence-plugin/blob/master/maven-confluence-reporting-plugin/src/main/java/org/bsc/maven/confluence/plugin/AbstractBaseConfluenceMojo.java
-        //
-        // FIX to resolve
-        // org.sonatype.plexus.components.sec.dispatcher.SecDispatcherException:
-        // java.io.FileNotFoundException: ~/.settings-security.xml (No such file or directory)
-        //
-        //CSON: LineLength
-        if (securityDispatcher instanceof DefaultSecDispatcher) {
-            ((DefaultSecDispatcher) securityDispatcher).setConfigurationFile("~/.m2/settings-security.xml");
-        }
-
-        return securityDispatcher.decrypt(server.getPassword());
     }
 
     /**
@@ -1838,7 +1597,7 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
         for (Dependency d : dependencies) {
             boolean addName = true;
             for (Vulnerability v : d.getVulnerabilities()) {
-                if (failBuildOnAnyVulnerability || v.getCvssV2Score() >= failBuildOnCVSS) {
+                if (failBuildOnAnyVulnerability || v.getCvssV2().getScore() >= failBuildOnCVSS) {
                     if (addName) {
                         addName = false;
                         ids.append(NEW_LINE).append(d.getFileName()).append(": ");
@@ -1877,62 +1636,8 @@ public abstract class BaseDependencyCheckMojo extends AbstractMojo implements Ma
      */
     protected void showSummary(MavenProject mp, Dependency[] dependencies) {
         if (showSummary) {
-            final StringBuilder summary = new StringBuilder();
-            for (Dependency d : dependencies) {
-                boolean firstEntry = true;
-                final StringBuilder ids = new StringBuilder();
-                for (Vulnerability v : d.getVulnerabilities()) {
-                    if (firstEntry) {
-                        firstEntry = false;
-                    } else {
-                        ids.append(", ");
-                    }
-                    ids.append(v.getName());
-                }
-                if (ids.length() > 0) {
-                    summary.append(d.getFileName()).append(" (");
-                    firstEntry = true;
-                    for (Identifier id : d.getIdentifiers()) {
-                        if (firstEntry) {
-                            firstEntry = false;
-                        } else {
-                            summary.append(", ");
-                        }
-                        summary.append(id.getValue());
-                    }
-                    summary.append(") : ").append(ids).append(NEW_LINE);
-                }
-            }
-            if (summary.length() > 0) {
-                final String msg = String.format("%n%n" + "One or more dependencies were identified with known vulnerabilities in %s:%n%n%s"
-                        + "%n%nSee the dependency-check report for more details.%n%n", mp.getName(), summary.toString());
-                getLog().warn(msg);
-            }
+            DependencyCheckScanAgent.showSummary(mp.getName(), dependencies);
         }
-    }
-
-    //</editor-fold>
-    //<editor-fold defaultstate="collapsed" desc="Methods to read/write the serialized data file">
-    /**
-     * Returns the key used to store the path to the data file that is saved by
-     * <code>writeDataFile()</code>. This key is used in the
-     * <code>MavenProject.(set|get)ContextValue</code>.
-     *
-     * @return the key used to store the path to the data file
-     */
-    protected String getDataFileContextKey() {
-        return "dependency-check-path-" + dataFileName;
-    }
-
-    /**
-     * Returns the key used to store the path to the output directory. When
-     * generating the report in the <code>executeAggregateReport()</code> the
-     * output directory should be obtained by using this key.
-     *
-     * @return the key used to store the path to the output directory
-     */
-    protected String getOutputDirectoryContextKey() {
-        return "dependency-output-dir-" + dataFileName;
     }
 
     //</editor-fold>